refactor: cleanup imports and duplicate code
