+++ package.json
@@ -102,7 +102,6 @@
     "@babel/parser": "^7.23.0",
     "@babel/plugin-syntax-import-assertions": "^7.22.5",
     "@babel/types": "^7.23.0",
-    "@intlify/vue-router-bridge": "^1.1.0",
     "@nuxt/module-builder": "^0.5.4",
     "@nuxt/schema": "^3.7.4",
     "@types/debug": "^4.1.9",
