+++ src/runtime/routing/compatibles/utils.ts
@@ -1,8 +1,8 @@
 import { useRouter } from '#imports'
 import { assign } from '@intlify/shared'
 
-import type { RouteLocationNormalizedLoaded, RouteLocationPathRaw } from 'vue-router'
 import type { Locale } from 'vue-i18n'
+import type { RouteLocationNormalizedLoaded, RouteLocationPathRaw } from 'vue-router'
 import type { Strategies } from '#build/i18n.options.mjs'
 
 function split(str: string, index: number) {
